add junit5

+++ src/main/java/com/jero/personal/utilities/repository/UserRepository.java
@@ -3,5 +3,8 @@ package com.jero.personal.utilities.repository;
 import com.jero.personal.utilities.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+/**
+ * The interface User repository.
+ */
+public interface UserRepository extends CrudRepository<User, Integer> {
 }
